Refactor BuildFailure code (#2250)
* Refactor BuildFailure items into corresponding kt file

* Add tests for IssueExtension
